Revert "Merge branch 'feature-custom-style' into main"
This reverts commit c5137f84881f1a85f66fa8c6e349e3fd0d23ebef, reversing
changes made to 26933198533fd6ce1d1df1f9c65fde5ca27fa47d.

+++ android/src/main/kotlin/dev/note11/flutter_naver_map/flutter_naver_map/model/map/overlay/overlay/NPolylineOverlay.kt
@@ -23,7 +23,7 @@ internal data class NPolylineOverlay(
     val widthDp: Double,
     val lineCap: LineCap,
     val lineJoin: LineJoin,
-    val patternDp: List<Double>,
+    val patternDp: List<Int>,
 ) : AddableOverlay<PolylineOverlay>() {
 
 
@@ -36,7 +36,7 @@ internal data class NPolylineOverlay(
         g.width = DisplayUtil.dpToPx(widthDp)
         g.capType = lineCap
         g.joinType = lineJoin
-        g.setPattern(*patternDp.map { DisplayUtil.dpToPx(it) }.toIntArray())
+        g.setPattern(*patternDp.map { DisplayUtil.dpToPx(it.toDouble()) }.toIntArray())
     }
 
     companion object {
@@ -48,7 +48,7 @@ internal data class NPolylineOverlay(
                 widthDp = it[widthName]!!.asDouble(),
                 lineCap = it[lineCapName]!!.asLineCap(),
                 lineJoin = it[lineJoinName]!!.asLineJoin(),
-                patternDp = it[patternName]!!.asList { p -> p.asDouble() },
+                patternDp = it[patternName]!!.asList { p -> p.asInt() },
             )
         }
